// frontend/src/app/api/model/assessmentGradeDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AssessmentGradeDto { 
    id?: string;
    assessmentElementId?: string;
    studentId?: string;
    grade?: number;
    feedback?: string | null;
    gradedByStaffId?: string;
    gradedAtUtc?: string;
}